+++ src/main/java/ladysnake/effective/client/EffectiveConfig.java
@@ -4,13 +4,13 @@ import eu.midnightdust.lib.config.MidnightConfig;
 
 public class EffectiveConfig extends MidnightConfig {
     @Entry
-    public static boolean generateSplashes = true;
+    public static boolean enableSplashes = true;
 
     @Entry(min = 0f, max = 1.0f)
     public static float splashRimAlpha = 0.7f;
 
     @Entry
-    public static boolean generateCascades = true;
+    public static boolean enableCascades = true;
 
     @Entry(min = 0, max = 500)
     public static int cascadeSoundsVolumeMultiplier = 100;
@@ -29,6 +29,9 @@ public class EffectiveConfig extends MidnightConfig {
 
     @Entry(min = 0, max = 10)
     public static int rainRippleDensity = 1;
+
+    @Entry
+    public static boolean enableGlowingPlankton = true;
 
     @Entry
     public static boolean glowsquidHypnotize = true;